+++ src/Model.h
@@ -33,12 +33,8 @@ unsigned int TextureFromFile(const char* path, const std::string& directory, boo
 
 struct BoneInfo
 {
-    /*id is index in finalBoneMatrices*/
     int id;
-
-    /*offset matrix transforms vertex from model space to bone space*/
     glm::mat4 offset;
-
 };
 
 class Model
@@ -62,314 +58,56 @@ public:
     std::map<std::string, BoneInfo> m_BoneInfoMap;
     int m_BoneCounter = 0;
 
+public:
     Model() {}
 
+    Model(std::string_view path_in, Shader& shader_in, LightingShaderRoutine& shaderRoutine_in, bool gamma = false, bool rotate_in = false, bool constructSubobjects_in = false); 
 
-    Model(std::string_view path_in, Shader& shader_in, LightingShaderRoutine& shaderRoutine_in, bool gamma = false, bool rotate_in = false, bool constructSubobjects_in = false) 
-        : shader(shader_in), shaderRoutine(shaderRoutine_in), gammaCorrection(gamma), rotate(rotate_in), path(path_in)
-    {
-		loadModel();
-    }
+    Model(Mesh mesh_in, Shader shader_in, LightingShaderRoutine shaderRoutine_in); 
 
-    Model(Mesh mesh_in, Shader shader_in, LightingShaderRoutine shaderRoutine_in) : shader(shader_in), shaderRoutine(shaderRoutine_in)
-    {
-        meshes.push_back(mesh_in);
-    }
+    Model(Shader shader_in, LightingShaderRoutine& shaderRoutine_in);
 
-    Model(Shader shader_in, LightingShaderRoutine& shaderRoutine_in) : shader(shader_in), shaderRoutine(shaderRoutine_in)
-    {
-    }
+    Model(std::string path_in, LightingShaderRoutine& sr, Shader shader_in, bool rotate_in = false);
 
-    Model(std::string path_in, LightingShaderRoutine& sr, Shader shader_in, bool rotate_in = false) : path(path_in), rotate(rotate_in), shaderRoutine(sr), shader(shader_in)
-    {
-    }
+    Model(const Model& m);
 
-    Model(const Model& m) : meshes(m.meshes), shader(m.shader), shaderRoutine(m.shaderRoutine)
-    {
-
-    }
-
-    Model(std::string_view path_in) : path(path_in){
-        loadModel();
-    }
+    Model(std::string_view path_in);
 
     //TODO(darius) make it shaderRoutine(tr[i]) for each mesh
-    void Draw(Transform tr, std::optional<PointLight>& light, std::optional<Material>& m)
-    {
-        if(light){
-            light->setShaderLight(shader);
-        }
-        if(m){
-            m->setShaderMaterial(shader);
-        }
+    void Draw(Transform tr, std::optional<PointLight>& light, std::optional<Material>& m);
 
-        if (animationShaderRoutine)
-            animationShaderRoutine->operator()(tr);
-        else
-			shaderRoutine(tr);
+    std::vector<Mesh> loadModel();
 
-        for (unsigned int i = 0; i < meshes.size(); i++) {
-            //if(i == 3)
-			//	shaderRoutine(Transform{(tr.position + glm::vec3{3,3,3}), tr.scale});
-            meshes[i].Draw(shader);
-        }
-    }
+    void addMesh(const Mesh& m);
 
-    std::vector<Mesh> loadModel()
-    {
-        if (path == "")
-            return {};
-        Assimp::Importer importer;
-        std::cout << "Read started...\n";
-        const aiScene* scene = importer.ReadFile(path.data(), aiProcess_Triangulate | aiProcess_GenSmoothNormals | aiProcess_FlipUVs | aiProcess_CalcTangentSpace);
-        std::cout << "Read ended!\n";
+    LightingShaderRoutine& getShaderRoutine();
 
-        if (!scene || scene->mFlags & AI_SCENE_FLAGS_INCOMPLETE || !scene->mRootNode) 
-        {
-            std::cout << "ERROR::ASSIMP:: " << importer.GetErrorString() << std::endl;
-            return {};
-        }
-       
-        directory = path.substr(0, path.find_last_of('/'));
+    void setAnimationShaderRoutine(SkeletalAnimationShaderRoutine r);
 
-        processNode(scene->mRootNode, scene);
-
-        return meshes;
-    }
-
-    void addMesh(const Mesh& m)
-    {
-        meshes.push_back(m);
-    }
-
-    LightingShaderRoutine& getShaderRoutine()
-    {
-        return shaderRoutine; 
-    }
-
-    void setAnimationShaderRoutine(SkeletalAnimationShaderRoutine r)
-    {
-        animationShaderRoutine = r;
-    }
-
-    void setShader(Shader sdr)
-    {
-        shader = sdr;
-    }
+    void setShader(Shader sdr);
 
     auto& GetBoneInfoMap() { return m_BoneInfoMap; }
-    int& GetBoneCount() { return m_BoneCounter; }    
+    int& GetBoneCount() { return m_BoneCounter; } 
 
-    void SetVertexBoneDataToDefault(Vertex& vertex)
-    {
-        for (int i = 0; i < MAX_BONE_INFLUENCE; i++)
-        {
-            vertex.m_BoneIDs[i] = -1;
-            vertex.m_Weights[i] = 0.0f;
-        }
-    }
+    void SetVertexBoneDataToDefault(Vertex& vertex);
 
-    void SetVertexBoneData(Vertex& vertex, int boneID, float weight)
-    {
-        for (int i = 0; i < MAX_BONE_INFLUENCE; ++i)
-        {
-            if (vertex.m_BoneIDs[i] < 0)
-            {
-                vertex.m_Weights[i] = weight;
-                vertex.m_BoneIDs[i] = boneID;
-                break;
-            }
-        }
-    }
+    void SetVertexBoneData(Vertex& vertex, int boneID, float weight);
 
 private:
     //TODO(darius) Bench it for recursive BFS, i think recursive is faster
     //TODO(darius) Dont use std::queue
-    void processNode(aiNode* root, const aiScene* scene)
-    {
-        std::queue<aiNode*> path;
+    void processNode(aiNode* root, const aiScene* scene);
 
-        path.push(root);
+    Mesh processMesh(aiMesh* mesh, const aiScene* scene);
 
-        while (!path.empty()) {
-            auto* curr = path.front();
-            path.pop();
+    std::vector<Texture> loadMaterialTextures(aiMaterial* mat, aiTextureType type, std::string typeName);
 
-            for (unsigned int i = 0; i < curr->mNumMeshes; ++i)
-            {
-                aiMesh* mesh = scene->mMeshes[curr->mMeshes[i]];
-                meshes.push_back(processMesh(mesh, scene));
-            }
- 
-            for (unsigned int i = 0; i < curr->mNumChildren; i++)
-			{
-                path.push(curr -> mChildren[i]);
-			}
-        }
-    }
-
-    Mesh processMesh(aiMesh* mesh, const aiScene* scene)
-    {
-        std::vector<Vertex> vertices;
-        std::vector<unsigned int> indices;
-        std::vector<Texture> textures;
-
-        for (unsigned int i = 0; i < mesh->mNumVertices; i++)
-        {
-            Vertex vertex;
-
-            SetVertexBoneDataToDefault(vertex);
-
-            glm::vec3 vector; 
-
-            vector.x = mesh->mVertices[i].x;
-            vector.y = mesh->mVertices[i].y;
-            vector.z = mesh->mVertices[i].z;
-            vertex.Position = vector;
-
-            if (mesh->HasNormals())
-            {
-                vector.x = mesh->mNormals[i].x;
-                vector.y = mesh->mNormals[i].y;
-                vector.z = mesh->mNormals[i].z;
-                vertex.Normal = vector;
-            }
-
-            if (mesh->mTextureCoords[0]) 
-            {
-                glm::vec2 vec;
-                vec.x = mesh->mTextureCoords[0][i].x;
-                vec.y = mesh->mTextureCoords[0][i].y;
-                vertex.TexCoords = vec;
-                // tangent
-                vector.x = mesh->mTangents[i].x;
-                vector.y = mesh->mTangents[i].y;
-                vector.z = mesh->mTangents[i].z;
-                vertex.Tangent = vector;
-                // bitangent
-                vector.x = mesh->mBitangents[i].x;
-                vector.y = mesh->mBitangents[i].y;
-                vector.z = mesh->mBitangents[i].z;
-                vertex.Bitangent = vector;
-            }
-            else
-                vertex.TexCoords = glm::vec2(0.0f, 0.0f);
-
-            vertices.push_back(vertex);
-        }
-
-        for (unsigned int i = 0; i < mesh->mNumFaces; i++)
-        {
-            aiFace face = mesh->mFaces[i];
-            for (unsigned int j = 0; j < face.mNumIndices; j++)
-                indices.push_back(face.mIndices[j]);
-        }
-
-        aiMaterial* material = scene->mMaterials[mesh->mMaterialIndex];
-
-        // 1. diffuse maps
-        std::vector<Texture> diffuseMaps = loadMaterialTextures(material, aiTextureType_DIFFUSE, "texture_diffuse");
-        textures.insert(textures.end(), diffuseMaps.begin(), diffuseMaps.end());
-        // 2. specular maps
-        std::vector<Texture> specularMaps = loadMaterialTextures(material, aiTextureType_SPECULAR, "texture_specular");
-        textures.insert(textures.end(), specularMaps.begin(), specularMaps.end());
-        // 3. normal maps
-        std::vector<Texture> normalMaps = loadMaterialTextures(material, aiTextureType_HEIGHT, "texture_normal");
-        textures.insert(textures.end(), normalMaps.begin(), normalMaps.end());
-        // 4. height maps
-        std::vector<Texture> heightMaps = loadMaterialTextures(material, aiTextureType_AMBIENT, "texture_height");
-        textures.insert(textures.end(), heightMaps.begin(), heightMaps.end());
-
-
-        ExtractBoneWeightForVertices(vertices,mesh,scene);
-
-        return Mesh(vertices, indices, textures);
-    }
-
-    std::vector<Texture> loadMaterialTextures(aiMaterial* mat, aiTextureType type, std::string typeName)
-    {
-        std::vector<Texture> textures;
-        for (unsigned int i = 0; i < mat->GetTextureCount(type); i++)
-        {
-            aiString str;
-
-            mat->GetTexture(type, i, &str);
-            std::cout << "textures: " << str.C_Str() << "\n";
-
-
-            auto fnd = Model::textures_set.find(str.C_Str());
-            if (fnd != Model::textures_set.end()) 
-            {
-                for (size_t j = 0; j < Model::textures_loaded.size(); ++j) {
-                    if (std::strcmp(Model::textures_loaded[j].get_path().data(), (*fnd).data()) == 0) {
-                        textures.push_back(Model::textures_loaded[j]);
-                    }
-                }
-            }
-            else 
-            {
-				Texture texture(TextureFromFile(str.C_Str(), this->directory, gammaCorrection, rotate), str.C_Str(), typeName);
-                textures.push_back(texture);
-                Model::textures_loaded.push_back(texture);
-                Model::textures_set.insert(str.C_Str());
-            }
-        }
-        return textures;
-    }
-
-    void ExtractBoneWeightForVertices(std::vector<Vertex>& vertices, aiMesh* mesh, const aiScene* scene)
-    {
-        for (int boneIndex = 0; boneIndex < mesh->mNumBones; ++boneIndex)
-        {
-            int boneID = -1;
-            std::string boneName = mesh->mBones[boneIndex]->mName.C_Str();
-            if (m_BoneInfoMap.find(boneName) == m_BoneInfoMap.end())
-            {
-                BoneInfo newBoneInfo;
-                newBoneInfo.id = m_BoneCounter;
-                newBoneInfo.offset = ConvertMatrixToGLMFormat(
-                    mesh->mBones[boneIndex]->mOffsetMatrix);
-                m_BoneInfoMap[boneName] = newBoneInfo;
-                boneID = m_BoneCounter;
-                m_BoneCounter++;
-            }
-            else
-            {
-                boneID = m_BoneInfoMap[boneName].id;
-            }
-            assert(boneID != -1);
-            auto weights = mesh->mBones[boneIndex]->mWeights;
-            int numWeights = mesh->mBones[boneIndex]->mNumWeights;
-
-            for (int weightIndex = 0; weightIndex < numWeights; ++weightIndex)
-            {
-                int vertexId = weights[weightIndex].mVertexId;
-                float weight = weights[weightIndex].mWeight;
-                assert(vertexId <= vertices.size());
-                SetVertexBoneData(vertices[vertexId], boneID, weight);
-            }
-        }
-    }
+    void ExtractBoneWeightForVertices(std::vector<Vertex>& vertices, aiMesh* mesh, const aiScene* scene);
+  
 public:
-    static inline glm::mat4 ConvertMatrixToGLMFormat(const aiMatrix4x4& from)
-    {
-        glm::mat4 to;
-        //the a,b,c,d in assimp is the row ; the 1,2,3,4 is the column
-        to[0][0] = from.a1; to[1][0] = from.a2; to[2][0] = from.a3; to[3][0] = from.a4;
-        to[0][1] = from.b1; to[1][1] = from.b2; to[2][1] = from.b3; to[3][1] = from.b4;
-        to[0][2] = from.c1; to[1][2] = from.c2; to[2][2] = from.c3; to[3][2] = from.c4;
-        to[0][3] = from.d1; to[1][3] = from.d2; to[2][3] = from.d3; to[3][3] = from.d4;
-        return to;
-    }
+    static glm::mat4 ConvertMatrixToGLMFormat(const aiMatrix4x4& from);
 
-    static inline glm::vec3 GetGLMVec(const aiVector3D& vec) 
-    { 
-        return glm::vec3(vec.x, vec.y, vec.z); 
-    }
+    static glm::vec3 GetGLMVec(const aiVector3D& vec); 
 
-    static inline glm::quat GetGLMQuat(const aiQuaternion& pOrientation)
-    {
-        return glm::quat(pOrientation.w, pOrientation.x, pOrientation.y, pOrientation.z);
-    }
+    static glm::quat GetGLMQuat(const aiQuaternion& pOrientation);
 };
-
