+++ src/SkeletalAnimationShaderRoutine.cpp
@@ -8,19 +8,22 @@ void SkeletalAnimationShaderRoutine::operator() (Transform tr){
 	    auto transforms = animator->GetFinalBoneMatrices();
 
         auto tr = glm::mat4(1.0f);
-        glm::mat4 ViewRotateX = glm::rotate(
+        /*glm::mat4 ViewRotateX = glm::rotate(
             tr,
             3.14f,
             glm::vec3(-1.0f, 0.0f, 0.0f)
         );
+        for(int i = 0; i < 99; ++i)
+			transforms[i] = ViewRotateX;
+            */
         for (int i = 0; i < transforms.size(); ++i)
-            sv.setMat4("finalBonesMatrices[" + std::to_string(i) + "]", ViewRotateX);
+            sv.setMat4("finalBonesMatrices[" + std::to_string(i) + "]", transforms[i]);
     }
 
     glm::mat4 model = glm::mat4(1.0f);
     glm::vec3 pos = tr.position;
     glm::mat4 q = tr.get_quatmat();
-    glm::vec3 scale = tr.scale;
+    glm::vec3 scale = { tr.scale.x/100, tr.scale.y/100, tr.scale.z/100 };
 
     model = glm::translate(model, pos);
     model *= q;
